+++ src/chatbot/main.py
@@ -1,32 +1,26 @@
 import json
 import os
-from datetime import datetime
 
-import models
+import services
 import uvicorn
 from dotenv import load_dotenv
-from fastapi import FastAPI, Request
+from fastapi import APIRouter, FastAPI, Request
 from fastapi.responses import JSONResponse
 from linebot import LineBotApi
 from linebot.exceptions import InvalidSignatureError, LineBotApiError
 from linebot.models import MessageEvent, TextMessage, TextSendMessage
 from linebot.v3.webhook import WebhookHandler
 
-import database
-
 app = FastAPI()
 
+router = APIRouter(prefix="/linebot")
 
-# Load environment variables
 load_dotenv()
 line_bot_api = LineBotApi(os.getenv("CHANNEL_ACCESS_TOKEN"))
 handler = WebhookHandler(os.getenv("CHANNEL_SECRET"))
-if line_bot_api is None or handler is None:
-    print("Make sure you have the right environment variables")
 
 
-# Deal with incoming messages
-@app.post("/callback")
+@router.post("/callback")
 async def line_webhook(request: Request):
     body_bytes = await request.body()
     body_str = body_bytes.decode("utf-8")
@@ -35,81 +29,46 @@ async def line_webhook(request: Request):
     try:
         handler.handle(body_str, signature)
     except InvalidSignatureError:
-        return JSONResponse(  # noqa
+        return JSONResponse(
             status_code=400, content={"message": "Invalid signature"}  # noqa
-        )
+        )  # noqa
 
     body = json.loads(body_str)
     events = body["events"]
 
     for event in events:
-        if (
-            event["source"]["type"] == "group"
-            and event["message"]["type"] == "text"  # noqa
-        ):
-            group_id = await group_chat(event)
-            # Print out the records from DB
-            print("-" * 40)
-            await output_group_msg(group_id)
-        elif (
-            event["source"]["type"] == "user"
-            and event["message"]["type"] == "text"  # noqa
-        ):
-            await user_chat(event)
-        else:
-            print("The message is not in form of text.")
+        event_type = event.get("type")
+        print(f"Processing event type: {event_type}")
+        if event_type == "message":
 
-    return JSONResponse(status_code=200, content={"message": "OK"})
+            if (
+                event["source"]["type"] == "group"
+                and event["message"]["type"] == "text"
+            ):
+                group_id = await services.group_chat(event)  # Use the function
+                print("-" * 40)
+                await services.output_group_msg(group_id)
+            elif (
+                event["source"]["type"] == "user"
+                and event["message"]["type"] == "text"  # noqa  # noqa
+            ):
+                await services.user_chat(event)
+            else:
+                print("The message is not in form of text.")
 
-
-# Deal with group messages
-async def group_chat(event):
-    message = event["message"]
-
-    if "groupId" in event["source"]:
-        group_id = event["source"]["groupId"]
-        user_id = event["source"]["userId"]
-
-        profile = line_bot_api.get_group_member_profile(group_id, user_id)
-        user_name = profile.display_name
-
-        # Build a MsgDetail instance to save the message
-        msg_detail = models.MsgDetail(
-            UserId=user_id,
-            UserName=user_name,
-            MsgText=message["text"],
-            Time=datetime.now(),
-        )
-        # Insert the message to MongoDB
-        try:
-            await database.add_message_to_group(group_id, msg_detail)
-            print(f"Insert message to {group_id} successfully.")
-        except Exception as e:
-            print(f"Exception: {e}")
-    return group_id
+        return JSONResponse(status_code=200, content={"message": "OK"})
+    else:
+        print(f"Unhandled event type: {event_type}")
 
 
-async def user_chat(events):
-    pass
-
-
-# output group chat messages
-async def output_group_msg(group_id):
-    messages = await database.select_all_group_msg(group_id)
-    for message in messages:
-        print("FROM MONGO DB:", message)
-
-
-# Webhook endpoint for dialogflow
-@app.post("/webhook")
+@router.post("/webhook")
 async def webhook(request: Request):
     req_data = await request.json()
-    # print out user info
     payload = req_data["originalDetectIntentRequest"]["payload"]
     user_id = payload["data"]["source"]["userId"]
     print(f"User ID: {user_id}")
     try:
-        profile = line_bot_api.get_profile(user_id)
+        profile = services.line_bot_api.get_profile(user_id)
         print(f"Profile: {profile}")
     except LineBotApiError as error:
         print(f"Error: {error}")
@@ -124,10 +83,12 @@ async def webhook(request: Request):
 
 @handler.add(MessageEvent, message=TextMessage)
 def handle_message(event):
-    line_bot_api.reply_message(
+    services.line_bot_api.reply_message(
         event.reply_token, TextSendMessage(text=event.message.text)
     )
 
+
+app.include_router(router)
 
 if __name__ == "__main__":
     uvicorn.run("main:app", host="0.0.0.0", port=8080, reload=True)
